# docs/fm-module/api.yaml
openapi: 3.1.0
info:
  title: WISLA FM — fm-module BFF API
  version: 1.0.0
  description: |
    REST API for WISLA Fault Management module (BFF on port 8080).
    Serves Angular SPA and adapter ingestion. JSON bodies use **camelCase**.
    PostgreSQL database: `wisla_fm`.
servers:
  - url: http://localhost:8080
    description: fm-module (BFF + static Angular)
...
paths:
...
  /api/v1/raw-events:
    get:
      tags: [raw-events]
      summary: List primary / pre-processing events
      operationId: listRawEvents
      x-mvp: true
      parameters:
        - name: sourceId
          in: query
          schema:
            type: string
            format: uuid
        - name: severity
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Severity'
          style: form
          explode: true
        - name: query
          in: query
          schema:
            type: string
        - name: processed
          in: query
          description: Filter by processing status
          schema:
            type: boolean
        - name: createdFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: createdTo
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Paged raw events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawEventPage'
...
  /api/v1/rules:
    get:
      tags: [rules]
      summary: List processing rules
      operationId: listRules
      x-mvp: true
      parameters:
        - name: ruleType
          in: query
          schema:
            $ref: '#/components/schemas/RuleType'
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: approvalStatus
          in: query
          schema:
            $ref: '#/components/schemas/ApprovalStatus'
      responses:
        '200':
          description: Rule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProcessingRule'
    post:
      tags: [rules]
      summary: Create processing rule
      description: |
        When `canvas` is present, the server validates graph structure before persist:
        at least one **stream trigger** (`type=trigger`, `data.triggerType=stream`) reachable
        from the graph entry, and at least one **action block** (`dedup`, `threshold`,
        `correlation`, `notify`, or `push`) on a path from that trigger. Orphan nodes (not connected to the
        trigger subgraph) are rejected. **notify** blocks require `data.channel` (`telegram` or `email`);
        when `channel=email`, `data.emailAddress` must be a valid email. `status` blocks are accepted but not executed at
        runtime (post-MVP). **notify** is a stub at runtime (no real SMTP/Telegram); **push** persists in-app notifications for UI polling.
      operationId: createRule
      x-mvp: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessingRuleCreate'
      responses:
        '201':
          description: Rule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingRuleDetail'
        '400':
          description: Invalid canvas structure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingStreamTrigger:
                  summary: No stream trigger
                  value:
                    error: canvas_validation
                    message: Rule canvas must include at least one trigger with triggerType=stream
                    details:
                      - field: canvas.nodes
                        message: missing_stream_trigger
                missingActionBlock:
                  summary: No action block on trigger path
                  value:
                    error: canvas_validation
                    message: Rule canvas must include at least one action block (dedup, threshold, correlation, notify, or push) reachable from the stream trigger
                    details:
                      - field: canvas.nodes
                        message: missing_action_block
                notifyMissingChannel:
                  summary: Notify block without channel
                  value:
                    error: canvas_validation
                    message: Notify block requires channel (telegram or email)
                    details:
                      - field: canvas.nodes
                        message: notify_missing_channel
                notifyInvalidEmail:
                  summary: Notify email channel without valid address
                  value:
                    error: canvas_validation
                    message: Notify block with email channel requires a valid emailAddress
                    details:
                      - field: canvas.nodes
                        message: notify_invalid_email
                orphanNodes:
                  summary: Disconnected nodes
                  value:
                    error: canvas_validation
                    message: Rule canvas contains nodes not reachable from the stream trigger
                    details:
                      - field: canvas.nodes
                        message: orphan_nodes
  /api/v1/rules/{id}:
    get:
      tags: [rules]
      summary: Get rule with canvas definition and versions
      operationId: getRule
      x-mvp: true
      parameters:
        - $ref: '#/components/parameters/ruleId'
      responses:
        '200':
          description: Rule detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingRuleDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [rules]
      summary: Update rule metadata or canvas
      description: |
        Same canvas validation as `POST /api/v1/rules` when `canvas` is included in the
        request body. Partial updates without `canvas` skip graph validation.
        When `enabled: true`, validates the stored (or patched) canvas if it has nodes;
        returns `400 canvas_validation` when the graph is invalid.
      operationId: patchRule
      x-mvp: true
      parameters:
        - $ref: '#/components/parameters/ruleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessingRulePatch'
      responses:
        '200':
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingRuleDetail'
        '400':
          description: Invalid canvas structure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [rules]
      summary: Delete rule
      operationId: deleteRule
      x-mvp: true
      parameters:
        - $ref: '#/components/parameters/ruleId'
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User session JWT from `/api/v1/auth/login`
    sourceApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Per-source API key for ingest
    serviceApiKey:
      type: apiKey
      in: header
      name: X-Service-Key
      description: Adapter service account key for internal endpoints
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
...
    ruleId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Severity:
      type: string
      enum: [fatal, critical, major, minor, warning, normal]
    EventStatus:
      type: string
      enum: [new, in_progress, closed, archived, maintenance, deferred]
...
    RuleType:
      type: string
      enum: [dedup, problem_resolution, correlation, threshold]
    ApprovalStatus:
      type: string
      enum: [approved, pending, draft]
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    PageMeta:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer

  # --- Auth ---
...
    Event:
      type: object
      required:
        - id
        - status
        - severity
        - title
        - description
        - repeatCount
        - ciId
        - nodeFqdn
        - systemName
        - sourceId
        - createdAt
        - sourceAt
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/EventStatus'
        severity:
          $ref: '#/components/schemas/Severity'
        title:
          type: string
        description:
          type: string
        assignedUserId:
          type: string
          format: uuid
        assignedUserName:
          type: string
          description: Denormalized display name (API only)
        repeatCount:
          type: integer
          minimum: 1
        ciId:
          type: string
          format: uuid
        nodeFqdn:
          type: string
        systemName:
          type: string
        subsystemName:
          type: string
        sourceId:
          type: string
          format: uuid
        sourceName:
          type: string
          description: Denormalized (API only)
        rootEventId:
          type: string
          format: uuid
          description: |
            Parent event ID when this event is a **child** in a correlation group.
            Set by the correlation canvas block when N matching events occur within the
            configured window; the earliest event in the window becomes the root.
            Omitted on root events and uncorrelated events.
        childEventIds:
          type: array
          items:
            type: string
            format: uuid
          description: |
            IDs of **child** events linked to this event as correlation root.
            Populated on `GET /api/v1/events/{id}` (EventDetail) when this event is the
            root of a correlation group. Empty or omitted when there are no children.
        itsmIncidentNumber:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        sourceAt:
          type: string
          format: date-time
        lastRepeatAt:
          type: string
          format: date-time
        takenAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RawEvent:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            isRaw:
              type: boolean
              default: true
            processed:
              type: boolean
            processedEventId:
              type: string
              format: uuid
            ingestBatchId:
              type: string
              format: uuid
    RawEventPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RawEvent'
        page:
          $ref: '#/components/schemas/PageMeta'

  # --- Adapters runtime ---
...
    ProcessingRule:
      type: object
      required:
        - id
        - name
        - ruleType
        - enabled
        - triggerType
        - approvalStatus
        - description
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        ruleType:
          $ref: '#/components/schemas/RuleType'
        enabled:
          type: boolean
        triggerType:
          type: string
          example: Событие потока
        lastRunAt:
          type: string
          format: date-time
        approvalStatus:
          $ref: '#/components/schemas/ApprovalStatus'
        description:
          type: string
    ProcessingRuleDetail:
      allOf:
        - $ref: '#/components/schemas/ProcessingRule'
        - type: object
          properties:
            canvas:
              $ref: '#/components/schemas/RuleCanvas'
            versions:
              type: array
              items:
                $ref: '#/components/schemas/RuleVersion'
    ProcessingRuleCreate:
      type: object
      required: [name, ruleType, triggerType]
      properties:
        name:
          type: string
        ruleType:
          $ref: '#/components/schemas/RuleType'
        triggerType:
          type: string
        description:
          type: string
        canvas:
          $ref: '#/components/schemas/RuleCanvas'
    ProcessingRulePatch:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        triggerType:
          type: string
        enabled:
          type: boolean
          description: Enable or disable rule. When set to true, canvas validation runs if the rule has nodes.
        canvas:
          $ref: '#/components/schemas/RuleCanvas'
    RuleBlockType:
      type: string
      enum: [trigger, condition, switch, dedup, threshold, correlation, notify, push, status]
      description: |
        Canvas node type. **Runtime-supported:** `trigger`, `condition`, `switch`, `dedup`,
        `threshold`, `correlation`, `notify`, `push`. `notify` records outbound intent only
        (stub — no real SMTP/Telegram in MVP+). `push` persists in-app notifications for UI polling.
        `status` is stored but not executed (post-MVP).
...
    RuleCanvasNode:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/RuleBlockType'
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
        data:
          type: object
          description: |
            Block label and type-specific config (mirrors frontend `RuleBlock.config`).

            | `type` | `data` keys | Notes |
            |--------|-------------|-------|
            | `trigger` | `triggerType` (`stream` \| `fm`), `label` | `stream` required for valid canvas |
            | `condition` | `field`, `operator`, `value`, `label` | `field` ∈ severity, title, status, nodeFqdn; `operator` ∈ eq, ne, contains |
            | `switch` | `label` | First matching outgoing branch continues traversal |
            | `dedup` | `key`, `label` | Optional merge key; default source+title+CI |
            | `threshold` | `count`, `windowMin`, `label` | Sliding window per source+CI |
            | `correlation` | `count`, `windowMin`, `matchField`, `label` | Links events as root/child when count reached in window |
            | `notify` | `channel` (`telegram` \| `email`), `emailAddress`, `label` | `emailAddress` required when `channel=email`; runtime stub (no real send in MVP+) |
            | `push` | `message`, `label` | In-app toast; `message` supports `{title}`, `{severity}` placeholders |
            | `status` | `status`, `label` | Post-MVP; not executed at runtime |
          additionalProperties: true
    RuleCanvasEdge:
      type: object
      required: [id, source, target]
      properties:
        id:
          type: string
        source:
          type: string
          description: Source node id
        target:
          type: string
          description: Target node id
        label:
          type: string
          description: Optional branch label for switch outgoing edges (default branch when omitted)
    RuleCanvas:
      type: object
      description: |
        Visual rule builder graph (Monq BP-style). Persisted on create/update; interpreted
        at runtime for enabled rules. Empty `nodes` falls back to legacy `ruleType` behavior.
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/RuleCanvasNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/RuleCanvasEdge'
    RuleVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        versionNumber:
          type: integer
        createdAt:
          type: string
          format: date-time
        createdByUserName:
          type: string
        comment:
          type: string
        canvas:
          $ref: '#/components/schemas/RuleCanvas'
...
security:
  - bearerAuth: []